+++ apis/v1alpha1/cluster_types.go
@@ -130,6 +130,13 @@ type CrdbClusterSpec struct {
 	// +operator-sdk:csv:customresourcedefinitions:type=spec,displayName="Cockroach Database Ingress"
 	// +optional
 	Ingress *IngressConfig `json:"ingress,omitempty"`
+	// (Optional) LogConfigMap define the config map which contains log configuration used to send the logs through the
+	// proper channels in the cockroachdb. Logging configuration is available for cockroach version v21.1.0 onwards.
+	// The logging configuration is taken in format of yaml file, you can check the logging configuration here (https://www.cockroachlabs.com/docs/stable/configure-logs.html#default-logging-configuration)
+	// The default logging for cockroach version v20.x or less is stderr, logging API is ignored for older versions.
+	// +operator-sdk:csv:customresourcedefinitions:type=spec,displayName="Cockroach Database Logging configuration config map"
+	// +optional
+	LogConfigMap string `json:"logConfigMap,omitempty"`
 }
 
 // +k8s:openapi-gen=true