+++ e2e/versionchecker/versionchecker_test.go
@@ -17,12 +17,15 @@ limitations under the License.
 package versionchecker_test
 
 import (
+	"encoding/json"
 	"testing"
 	"time"
 
 	corev1 "k8s.io/api/core/v1"
 	apiresource "k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	crdbv1alpha1 "github.com/cockroachdb/cockroach-operator/apis/v1alpha1"
 	"github.com/cockroachdb/cockroach-operator/e2e"
 	"github.com/cockroachdb/cockroach-operator/pkg/controller"
 	"github.com/cockroachdb/cockroach-operator/pkg/testutil"
@@ -65,6 +68,50 @@ func TestVersionCheckerJobPodPending(t *testing.T) {
 			Test: func(t *testing.T) {
 				require.NoError(t, sb.Create(builder.Cr()))
 				testutil.RequireAtMostOneVersionCheckerJob(t, sb, 300*time.Second)
+			},
+		},
+	}
+	steps.Run(t)
+}
+
+func TestLoggingAPIValidCheck(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping test in short mode.")
+	}
+	testLog := zapr.NewLogger(zaptest.NewLogger(t))
+
+	e := testenv.CreateActiveEnvForTest()
+	env := e.Start()
+	defer e.Stop()
+
+	sb := testenv.NewDiffingSandbox(t, env)
+	sb.StartManager(t, controller.InitClusterReconcilerWithLogger(testLog))
+
+	logJson := []byte(`{"sinkss": "invalidvalue"}`)
+	logConfig := make(map[string]interface{})
+	require.NoError(t, json.Unmarshal(logJson, &logConfig))
+	testutil.RequireLoggingConfigMap(t, sb, "logging-configmap", string(logJson))
+
+	builder := testutil.NewBuilder("crdb").Namespaced(sb.Namespace).WithNodeCount(3).WithTLS().
+		WithImage("cockroachdb/cockroach:v21.1.0").
+		WithPVDataStore("32Mi", "standard" /* default storage class in KIND */).
+		WithClusterLogging("logging-configmap")
+
+	steps := testutil.Steps{
+		{
+			Name: "validate the logging API input",
+			Test: func(t *testing.T) {
+				require.NoError(t, sb.Create(builder.Cr()))
+				crdbCluster := &crdbv1alpha1.CrdbCluster{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      builder.Cr().Name,
+						Namespace: builder.Cr().Namespace,
+					},
+				}
+				// Added sleep to reconcile the CR.
+				time.Sleep(2 * time.Second)
+				require.NoError(t, sb.Get(crdbCluster))
+				require.Equal(t, crdbCluster.Status.ClusterStatus, "Failed")
 			},
 		},
 	}