+++ pkg/actor/validate_version.go
@@ -21,6 +21,7 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"os/exec"
 	"strings"
 	"time"
 
@@ -59,6 +60,30 @@ func (v *versionChecker) GetActionType() api.ActionType {
 }
 
 func (v *versionChecker) Act(ctx context.Context, cluster *resource.Cluster, log logr.Logger) error {
+	log.V(DEBUGLEVEL).Info("starting to check the logging config provided")
+	//we refresh the resource to make sure we use the latest version
+	fetcher := resource.NewKubeFetcher(ctx, cluster.Namespace(), v.client)
+
+	if cluster.IsLoggingAPIEnabled() {
+		if logConfig, err := cluster.LoggingConfiguration(fetcher); err == nil {
+			log.V(DEBUGLEVEL).Info(fmt.Sprintf("Log configuration for the cockroach cluster: %s", logConfig))
+			var stderr bytes.Buffer
+			cmd := exec.Command("bash", "-c", fmt.Sprintf("cockroach debug check-log-config --log=%s", logConfig))
+			cmd.Stderr = &stderr
+			cErr := cmd.Run()
+			if cErr != nil || stderr.String() != "" {
+				log.Error(cErr, "The cockroachdb logging API is set to value that is not supported by the operator, See the default logging configuration here (https://www.cockroachlabs.com/docs/stable/configure-logs.html#default-logging-configuration) ")
+				return errors.New(stderr.String())
+			} else {
+				log.V(DEBUGLEVEL).Info("Validated the logging config")
+			}
+		} else {
+			vErr := ValidationError{Err: err}
+			log.Error(vErr, "The cockroachdb logging API value is set to a value that is not supported by the operator")
+			return err
+		}
+	}
+
 	log.V(DEBUGLEVEL).Info("starting to check the crdb version of the container provided")
 
 	r := resource.NewManagedKubeResource(ctx, v.client, cluster, kube.AnnotatingPersister)
@@ -272,6 +297,7 @@ func (v *versionChecker) Act(ctx context.Context, cluster *resource.Cluster, log
 		}
 
 		refreshedCluster := resource.NewCluster(cr)
+		refreshedCluster.Fetcher = fetcher
 		refreshedCluster.SetClusterVersion(calVersion)
 		refreshedCluster.SetAnnotationVersion(calVersion)
 		refreshedCluster.SetCrdbContainerImage(containerImage)
@@ -312,6 +338,7 @@ func (v *versionChecker) completeVersionChecker(
 	}
 
 	refreshedCluster := resource.NewCluster(cr)
+	refreshedCluster.Fetcher = fetcher
 	// save the status of the cluster
 	refreshedCluster.SetTrue(api.CrdbVersionChecked)
 	refreshedCluster.SetClusterVersion(version)